+++ src/phoenix/server/api/queries.py
@@ -80,11 +80,6 @@ from phoenix.server.api.types.pagination import (
 from phoenix.server.api.types.Project import Project
 from phoenix.server.api.types.ProjectSession import ProjectSession, to_gql_project_session
 from phoenix.server.api.types.Prompt import Prompt
-from phoenix.server.api.types.PromptVersion import (
-    PromptTemplateFormat,
-    PromptTemplateType,
-    PromptVersion,
-)
 from phoenix.server.api.types.SortDir import SortDir
 from phoenix.server.api.types.Span import Span, to_gql_span
 from phoenix.server.api.types.SystemApiKey import SystemApiKey
@@ -616,85 +611,6 @@ class Query:
             ],
             args=args,
         )
-
-    @strawberry.field
-    async def prompt_versions(
-        self,
-        info: Info[Context, None],
-        first: Optional[int] = 50,
-        last: Optional[int] = UNSET,
-        after: Optional[CursorString] = UNSET,
-        before: Optional[CursorString] = UNSET,
-    ) -> Connection[PromptVersion]:
-        args = ConnectionArgs(
-            first=first,
-            after=after if isinstance(after, CursorString) else None,
-            last=last,
-            before=before if isinstance(before, CursorString) else None,
-        )
-
-        dummy_data = [
-            PromptVersion(
-                id_attr=2,
-                user="alice",
-                description="A dummy prompt version",
-                template_type=PromptTemplateType.TEXT,
-                template_format=PromptTemplateFormat.MUSTACHE,
-                template={
-                    "_version": "messages-v1",
-                    "messages": [
-                        {"role": "user", "content": "Hello what's the weather in Antarctica like?"}
-                    ],
-                },
-                invocation_parameters={"temperature": 0.5},
-                tools={
-                    "_version": "tools-v1",
-                    "tools": [
-                        {
-                            "definition": {
-                                "name": "get_current_weather",
-                                "description": "Get the current weather in a given location",
-                                "parameters": {
-                                    "type": "object",
-                                    "properties": {
-                                        "location": {
-                                            "type": "string",
-                                            "description": "A location in the world",
-                                        },
-                                        "unit": {
-                                            "type": "string",
-                                            "enum": ["celsius", "fahrenheit"],
-                                            "default": "fahrenheit",
-                                            "description": "The unit of temperature",
-                                        },
-                                    },
-                                    "required": ["location"],
-                                },
-                            }
-                        }
-                    ],
-                },
-                model_name="gpt-4o",
-                model_provider="openai",
-            ),
-            PromptVersion(
-                id_attr=1,
-                user="alice",
-                description="A dummy prompt version",
-                template_type=PromptTemplateType.TEXT,
-                template_format=PromptTemplateFormat.MUSTACHE,
-                template={
-                    "_version": "messages-v1",
-                    "messages": [
-                        {"role": "user", "content": "Hello what's the weather in Antarctica like?"}
-                    ],
-                },
-                model_name="gpt-4o",
-                model_provider="openai",
-            ),
-        ]
-
-        return connection_from_list(data=dummy_data, args=args)
 
     @strawberry.field
     def clusters(
